Fixed many 64bit problems, including: write_de() took the value to write as a uint32_t, and only had a 4 byte buffer.
dact_common: Was not lseek()'ing properly for the new stream sizes.

net.c: Made lseek_net() allow forward seeking in streams

FossilOrigin-Name: ad581b297ec7ee6c532c693a5fcf7ce3e0a768e6

+++ dendian.h
@@ -21,7 +21,7 @@
 #ifndef __ENDIAN_LOCAL_H
 #define __ENDIAN_LOCAL_H
 
-int write_de(const int dst, const uint32_t num, const int sze);
+int write_de(const int dst, const uint64_t num, const int sze);
 int read_de(const int src, void *dest, const int sze, const int out_sze);
 
 /*
